Do not recirculate the lb prefix messages in tx_node
tx_node is too late for recirculation so do it before.
