+++ include/dp_util.h
@@ -6,6 +6,7 @@ extern "C" {
 #endif
 #include <stdbool.h>
 #include <rte_log.h>
+#include <rte_mbuf.h>
 
 #define DP_TIMESTAMP_BUF_SIZE 26
 
@@ -62,7 +63,7 @@ int get_op_env();
 int dp_is_wcmp_enabled();
 
 double dp_get_wcmp_frac();
-
+void rewrite_eth_hdr(struct rte_mbuf *m, uint16_t port_id, uint16_t eth_type);
 void print_ip(unsigned int ip, char *buf);
